weather_project/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0.3
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
packaging==21.3
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2022.1
redis==4.2.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.0
